read and create work, now I have to implement all of that into GET/POST requests
