+++ src/main/webapp/WEB-INF/includes/cartbox.jsp
@@ -1,3 +1,4 @@
+
 <!-- cart box -->
                 <div class="aa-cartbox">
                   <a class="aa-cart-link" href="#">
